requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
newspaper3k
schedule
